removing possible crashing causes

+++ PrYv-UX-PoC/NotesApp/AddNumericalValueViewController.m
@@ -275,7 +275,7 @@
         NSInteger selectedGroup = [_typePicker selectedRowInComponent:0];
         PYMeasurementTypesGroup *group = [_measurementGroups objectAtIndex:selectedGroup];
         
-        PYEventType *pyType = [group pyTypeAtIndex:row];
+        PYEventType *pyType = [group pyTypeAtIndex:(int)row];
         NSString *symbolText = pyType.type;
         
         NSString *nameText = @"";
